+++ app/src/main/java/com/example/android_chatbot/ui/chat_screen/ChatScreen.kt
@@ -34,7 +34,7 @@ fun ChatScreen(
     channelDAO: ChannelDAO,
     messageDAO: MessageDAO,
     settingDAO: SettingDAO,
-    channelId: Int,
+    channelId: Long,
     modifier: Modifier = Modifier
 ) {
     val viewModel: ChatViewModel =
